Discriminate pieces based on the player they belong to
I added an empty "config.py" file. It corresponds to a module that is imported in both board and score_board and that contains the turn variable (0 if turn of player 1 (black stones) and 1 if turn of player 2 (white stones)).
The color of the pieces on the board depends on the player to whom they belong.